Refactor: Consolidate structure - KISS principle
Simplifies project structure by consolidating all modules into modules/ directory.

Changes:
1. Moved views/reschedule_view.py → modules/reschedule_view.py
2. Moved pending_reschedules from shared_states.py into reschedule.py
3. Deleted modules/shared_states.py (no longer needed)
4. Deleted views/ directory (empty)

Updated imports:
- modules/reschedule.py:
  * Import reschedule_view from modules instead of views
  * Define pending_reschedules directly (no import)

- modules/reschedule_view.py:
  * Runtime import of pending_reschedules to avoid circular dependency
  * Import from modules.reschedule at usage points

- modules/admin_tools.py:
  * Import pending_reschedules from modules.reschedule

Benefits:
- KISS: Everything in one place (modules/)
- No separate views/ or shared_states.py needed
- Simpler directory structure
- Easier to find files
- Consistent with existing module organization

Structure before:
modules/
  reschedule.py
  shared_states.py (global state)
views/
  reschedule_view.py (UI components)

Structure after:
modules/
  reschedule.py (logic + global state)
  reschedule_view.py (UI components)
