--- modules/shared_states.py
@@ -1,4 +1,0 @@
-# modules/shared_states.py
-
-# Set of pending reschedule requests
-pending_reschedules = set()
